Collect Series.append tests (#32254)
